Move error handling in a separate class (#4597)
