+++ src/platforms/android/androidiaphandler.cpp
@@ -323,7 +323,7 @@ void AndroidIAPHandler::validatePurchase(QJsonObject purchase) {
       purchaseTask, &TaskPurchase::failed, this,
       [this](QNetworkReply::NetworkError error, const QByteArray&) {
         logger.error() << "Purchase validation request to guardian failed";
-        MozillaVPN::instance()->errorHandle(ErrorHandler::toErrorType(error));
+        ErrorHandler::instance()->errorHandle(ErrorHandler::toErrorType(error));
         stopSubscription();
         emit subscriptionNotValidated();
       });
