+++ src/platforms/ios/iosiaphandler.mm
@@ -339,7 +339,7 @@ void IOSIAPHandler::processCompletedTransactions(const QStringList& ids) {
 
             QJsonDocument json = QJsonDocument::fromJson(data);
             if (!json.isObject()) {
-              MozillaVPN::instance()->errorHandle(ErrorHandler::toErrorType(error));
+              ErrorHandler::instance()->errorHandle(ErrorHandler::toErrorType(error));
               emit subscriptionFailed();
               ErrorHandler::instance()->subscriptionGenericError();
               return;
@@ -348,7 +348,7 @@ void IOSIAPHandler::processCompletedTransactions(const QStringList& ids) {
             QJsonObject obj = json.object();
             QJsonValue errorValue = obj.value("errno");
             if (!errorValue.isDouble()) {
-              MozillaVPN::instance()->errorHandle(ErrorHandler::toErrorType(error));
+              ErrorHandler::instance()->errorHandle(ErrorHandler::toErrorType(error));
               emit subscriptionFailed();
               ErrorHandler::instance()->subscriptionGenericError();
               return;
@@ -356,14 +356,14 @@ void IOSIAPHandler::processCompletedTransactions(const QStringList& ids) {
 
             int errorNumber = errorValue.toInt();
             if (errorNumber == GUARDIAN_ERROR_RECEIPT_NOT_VALID) {
-              MozillaVPN::instance()->errorHandle(ErrorHandler::toErrorType(error));
+              ErrorHandler::instance()->errorHandle(ErrorHandler::toErrorType(error));
               emit subscriptionFailed();
               ErrorHandler::instance()->subscriptionExpiredError();
               return;
             }
 
             if (errorNumber == GUARDIAN_ERROR_RECEIPT_IN_USE) {
-              MozillaVPN::instance()->errorHandle(ErrorHandler::toErrorType(error));
+              ErrorHandler::instance()->errorHandle(ErrorHandler::toErrorType(error));
               emit subscriptionFailed();
               ErrorHandler::instance()->subscriptionInUseError();
               return;
