+++ src/platforms/linux/linuxcontroller.cpp
@@ -12,7 +12,6 @@
 #include "models/device.h"
 #include "models/keys.h"
 #include "models/server.h"
-#include "mozillavpn.h"
 
 #include <QDBusPendingCallWatcher>
 #include <QJsonDocument>
@@ -98,7 +97,7 @@ void LinuxController::operationCompleted(QDBusPendingCallWatcher* call) {
   QDBusPendingReply<bool> reply = *call;
   if (reply.isError()) {
     logger.error() << "Error received from the DBus service";
-    MozillaVPN::instance()->errorHandle(ErrorHandler::ControllerError);
+    ErrorHandler::instance()->errorHandle(ErrorHandler::ControllerError);
     emit disconnected();
     return;
   }
@@ -111,7 +110,7 @@ void LinuxController::operationCompleted(QDBusPendingCallWatcher* call) {
   }
 
   logger.error() << "DBus service says: error.";
-  MozillaVPN::instance()->errorHandle(ErrorHandler::ControllerError);
+  ErrorHandler::instance()->errorHandle(ErrorHandler::ControllerError);
   emit disconnected();
 }
