+++ src/ui/authenticationInApp/ViewAuthenticationUnblockCodeNeeded.qml
@@ -49,7 +49,7 @@ VPNInAppAuthenticationBase {
             anchors.horizontalCenter: parent.horizontalCenter
             onClicked: {
                 VPNAuthInApp.resendUnblockCodeEmail();
-                VPN.setAlert(VPN.AuthCodeSentAlert);
+                VPNErrorHandler.setAlert(VPNErrorHandler.AuthCodeSentAlert);
             
             }
         }
