+++ src/update/balrog.cpp
@@ -503,15 +503,12 @@ void Balrog::propagateError(NetworkRequest* request,
                             QNetworkReply::NetworkError error) {
   Q_ASSERT(request);
 
-  MozillaVPN* vpn = MozillaVPN::instance();
-  Q_ASSERT(vpn);
-
   // 451 Unavailable For Legal Reasons
   if (request->statusCode() == 451) {
     logger.debug() << "Geo IP restriction detected";
-    vpn->errorHandle(ErrorHandler::GeoIpRestrictionError);
+    ErrorHandler::instance()->errorHandle(ErrorHandler::GeoIpRestrictionError);
     return;
   }
 
-  vpn->errorHandle(ErrorHandler::toErrorType(error));
+  ErrorHandler::instance()->errorHandle(ErrorHandler::toErrorType(error));
 }
