+++ tests/auth/CMakeLists.txt
@@ -46,6 +46,9 @@ target_sources(auth_tests PRIVATE
     ${MVPN_SOURCE_DIR}/authenticationlistener.h
     ${MVPN_SOURCE_DIR}/constants.cpp
     ${MVPN_SOURCE_DIR}/constants.h
+    ${MVPN_SOURCE_DIR}/controller.h
+    ${MVPN_SOURCE_DIR}/dnspingsender.cpp
+    ${MVPN_SOURCE_DIR}/dnspingsender.h
     ${MVPN_SOURCE_DIR}/env.h
     ${MVPN_SOURCE_DIR}/errorhandler.cpp
     ${MVPN_SOURCE_DIR}/errorhandler.h
@@ -73,6 +76,14 @@ target_sources(auth_tests PRIVATE
     ${MVPN_SOURCE_DIR}/networkmanager.h
     ${MVPN_SOURCE_DIR}/networkrequest.cpp
     ${MVPN_SOURCE_DIR}/networkrequest.h
+    ${MVPN_SOURCE_DIR}/pinghelper.cpp
+    ${MVPN_SOURCE_DIR}/pinghelper.h
+    ${MVPN_SOURCE_DIR}/pingsender.cpp
+    ${MVPN_SOURCE_DIR}/pingsender.h
+    ${MVPN_SOURCE_DIR}/pingsenderfactory.cpp
+    ${MVPN_SOURCE_DIR}/pingsenderfactory.h
+    ${MVPN_SOURCE_DIR}/platforms/dummy/dummypingsender.cpp
+    ${MVPN_SOURCE_DIR}/platforms/dummy/dummypingsender.h
     ${MVPN_SOURCE_DIR}/rfc/rfc1918.cpp
     ${MVPN_SOURCE_DIR}/rfc/rfc1918.h
     ${MVPN_SOURCE_DIR}/rfc/rfc4193.cpp
@@ -126,6 +137,7 @@ target_sources(auth_tests PRIVATE
     testsignupandin.cpp
     testsignupandin.h
     mocmozillavpn.cpp
+    ../qml/moccontroller.cpp
     ../unit/mocinspectorhandler.cpp
 )
