+++ tests/auth/mocmozillavpn.cpp
@@ -23,6 +23,8 @@ MozillaVPN::MozillaVPN() {}
 
 MozillaVPN::~MozillaVPN() {}
 
+ConnectionHealth* MozillaVPN::connectionHealth() { return nullptr; }
+
 Controller* MozillaVPN::controller() { return nullptr; }
 
 MozillaVPN::State MozillaVPN::state() const { return StateInitialize; }
@@ -61,10 +63,6 @@ void MozillaVPN::accountChecked(const QByteArray&) {}
 void MozillaVPN::cancelAuthentication() {}
 
 void MozillaVPN::logout() {}
-
-void MozillaVPN::setAlert(AlertType) {}
-
-void MozillaVPN::errorHandle(ErrorHandler::ErrorType) {}
 
 const QList<Server> MozillaVPN::exitServers() const { return QList<Server>(); }
