+++ tests/qml/mocmozillavpn.cpp
@@ -25,6 +25,8 @@ MozillaVPN::MozillaVPN() {}
 
 MozillaVPN::~MozillaVPN() {}
 
+ConnectionHealth* MozillaVPN::connectionHealth() { return nullptr; }
+
 Controller* MozillaVPN::controller() { return new Controller(); }
 
 MozillaVPN::State MozillaVPN::state() const { return StateInitialize; }
@@ -67,10 +69,6 @@ void MozillaVPN::accountChecked(const QByteArray&) {}
 void MozillaVPN::cancelAuthentication() {}
 
 void MozillaVPN::logout() {}
-
-void MozillaVPN::setAlert(AlertType) {}
-
-void MozillaVPN::errorHandle(ErrorHandler::ErrorType) {}
 
 const QList<Server> MozillaVPN::exitServers() const { return QList<Server>(); }
