+++ tests/unit/mocmozillavpn.cpp
@@ -34,6 +34,8 @@ MozillaVPN::UserState MozillaVPN::userState() const {
 bool MozillaVPN::stagingMode() const { return true; }
 bool MozillaVPN::debugMode() const { return true; }
 
+ConnectionHealth* MozillaVPN::connectionHealth() { return nullptr; }
+
 Controller* MozillaVPN::controller() { return new Controller(); }
 
 void MozillaVPN::initialize() {}
@@ -66,10 +68,6 @@ void MozillaVPN::accountChecked(const QByteArray&) {}
 void MozillaVPN::cancelAuthentication() {}
 
 void MozillaVPN::logout() {}
-
-void MozillaVPN::setAlert(AlertType) {}
-
-void MozillaVPN::errorHandle(ErrorHandler::ErrorType) {}
 
 const QList<Server> MozillaVPN::exitServers() const { return QList<Server>(); }
